# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Market Spike API – gamified investing for kids and teens
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /friends/add:
    post:
      operationId: addFriend
      tags: [users]
      summary: Add a friend by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFriendRequest"
      responses:
        "200":
          description: New friend entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendEntry"
  /friends/{id}:
    delete:
      operationId: removeFriend
      tags: [users]
      summary: Remove a friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /market/stocks:
    get:
      operationId: listStocks
      tags: [market]
      summary: List all stocks with current prices
      parameters:
        - name: sector
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stock"
  /market/stocks/{symbol}:
    get:
      operationId: getStock
      tags: [market]
      summary: Get stock detail with price history
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockDetail"
...
  /learn/lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [learn]
      summary: Mark a lesson as complete and award XP
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonCompletion"
...
components:
  schemas:
...
    FriendEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarColor:
          type: string
        avatarUrl:
          type: ["string", "null"]
        xp:
          type: integer
        level:
          type: integer
        status:
          type: string
          enum: [pending, accepted]
        direction:
          type: string
          enum: [sent, received]
      required: [id, userId, username, avatarColor, xp, level, status, direction]
    AddFriendRequest:
      type: object
      properties:
        username:
          type: string
      required: [username]
...
    Stock:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        change:
          type: number
          description: Price change today (dollars)
        changePercent:
          type: number
          description: Price change today (percent)
        sector:
          type: string
        logoUrl:
          type: ["string", "null"]
        marketCap:
          type: ["string", "null"]
          description: Human-readable market cap (e.g. "$2.5T")
      required: [id, symbol, name, price, change, changePercent, sector]
    StockDetail:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        sector:
          type: string
        logoUrl:
          type: ["string", "null"]
        marketCap:
          type: ["string", "null"]
        description:
          type: string
        priceHistory:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
      required: [id, symbol, name, price, change, changePercent, sector, description, priceHistory]
    PricePoint:
      type: object
      properties:
        date:
          type: string
        price:
          type: number
      required: [date, price]
...
    LessonCompletion:
      type: object
      properties:
        xpEarned:
          type: integer
        newLevel:
          type: integer
        leveledUp:
          type: boolean
        newAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required: [xpEarned, newLevel, leveledUp, newAchievements]
    Achievement:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        xpReward:
          type: integer
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
      required: [id, name, description, icon, xpReward, rarity]
...